SoCOT/requirements.txt
absl-py==2.3.1
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
av==16.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
ctranslate2==4.6.3
cycler==0.12.1
decorator==4.4.2
deepface==0.0.97
faster-whisper==1.2.1
ffmpeg==1.4
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
jax==0.9.0
jaxlib==0.9.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.1
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightecc==0.0.4
lightphe==0.0.20
llvmlite==0.46.0
logging-config==1.0.4
lz4==4.4.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
ml_dtypes==0.5.4
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
namex==0.1.0
numba==0.63.1
numpy==2.3.5
onnxruntime==1.23.2
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==3.0.0
pillow==11.3.0
platformdirs==4.5.1
pooch==1.8.2
praat-parselmouth==0.4.7
proglog==0.1.12
protobuf==6.33.4
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.0.1
